Add indexed fold functions (#10)
* Add indexed fold functions

Add indexedFoldl and indexedFoldr. These functions are identical to
List.foldl and List.foldr except that the index of the current
element being processed is passed to the reducer function along with
the element itself. This functionality can be seen in JavaScript's
reduce and reduceRight.

* Update indexed fold functions

Adding indexedFoldl and indexedFoldr to the module exports,
switching the step function argument order (index first like
List.indexedMap) and general cleanup.