rewrite groupWhile to be tail-recursive
I have a very long list for which groupWhile blows the stack! Oh no! It looks like the accumulator function itself was tail-recursive, but not one of the functions it called. This unifies them and adds a bunch of comments to explain what's going on for the next person who needs to look at this code.